+++ CHANGES.md
@@ -6,6 +6,7 @@ This is a history of changes to [Remix](https://remix.run).
 
 ### Features
 
+- Added support for `headers({ loaderHeaders, parentsHeaders })` function in route modules
 - Added support for array initializer in `Headers`
 
 ### Bug Fixes